重改光线

+++ GlMainApp/GlMainApp/Base.cpp
@@ -51,13 +51,13 @@ void gdraw::drawTeaPot(float size, float x, float y, float z)
 void gdraw::print(float x, float y, float z, const char * str)
 {
 	glRasterPos3f(x, y, z);
-	for (int i = 0; i < strlen(str); i++)
+	for (size_t i = 0; i < strlen(str); i++)
 		glutBitmapCharacter(GLUT_BITMAP_HELVETICA_12, *(str + i));
 }
 
 void gdraw::wprint(float x, float y, float z, const wchar_t *str)
 {
 	glRasterPos3f(x, y, z);
-	for (int i = 0; i < wcslen(str); i++)
+	for (size_t i = 0; i < wcslen(str); i++)
 		glutBitmapCharacter(GLUT_BITMAP_HELVETICA_12, *(str + i));
 }
